pyproject.toml
[project]
name = "pydocs-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "annotated-types==0.5.0",
    "autocast==0.0.1b1",
    "beautifulsoup4==4.12.2",
    "blis==0.7.11",
    "catalogue==2.0.10",
    "cbor==1.0.0",
    "cbor2==5.5.0",
    "certifi==2023.7.22",
    "charset-normalizer==3.3.0",
    "click==8.1.7",
    "cloudpathlib==0.15.1",
    "coloredlogs==15.0.1",
    "confection==0.1.3",
    "contourpy==1.1.1",
    "cramjam==2.7.0",
    "cycler==0.12.1",
    "cymem==2.0.8",
    "cython==3.0.3",
    "faiss-cpu==1.7.4",
    "fastparquet==2023.8.0",
    "filelock==3.12.4",
    "flatbuffers==23.5.26",
    "fonttools==4.43.1",
    "fsspec==2023.9.2",
    "huggingface-hub==0.16.4",
    "humanfriendly==10.0",
    "idna==3.4",
    "ijson==3.2.3",
    "importlib-metadata==6.8.0",
    "importlib-resources==6.1.0",
    "inscriptis==2.3.2",
    "ir-datasets==0.5.5",
    "jinja2==3.1.2",
    "joblib==1.3.2",
    "kiwisolver==1.4.5",
    "langcodes==3.3.0",
    "lightgbm==4.1.0",
    "llvmlite==0.41.1",
    "lxml==4.9.3",
    "lz4==4.3.2",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.3",
    "matplotlib==3.7.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.10",
    "networkx==3.1",
    "nmslib==2.1.1",
    "numba==0.58.1",
    "numpy==1.26.0",
    "onnxruntime==1.16.0",
    "orjson==3.9.9",
    "packaging==23.2",
    "pandas==2.0.3",
    "pathy==0.10.2",
    "pillow==10.1.0",
    "pip==23.2.1",
    "preshed==3.0.9",
    "protobuf==4.24.4",
    "psutil==5.9.5",
    "pyautocorpus==0.1.12",
    "pydantic==2.4.2",
    "pydantic-core==2.10.1",
    "pygments==2.16.1",
    "pyjnius==1.5.0",
    "pyparsing==3.1.1",
    "pyserini==0.22.0",
    "python-dateutil==2.8.2",
    "pytrec-eval==0.5",
    "pytz==2023.3.post1",
    "pyyaml==6.0.1",
    "ranx==0.3.18",
    "regex==2023.10.3",
    "requests==2.31.0",
    "rich==13.6.0",
    "safetensors==0.3.3",
    "scikit-learn==1.3.1",
    "scipy==1.10.1",
    "seaborn==0.13.0",
    "sentencepiece==0.1.99",
    "setuptools==68.0.0",
    "six==1.16.0",
    "smart-open==6.4.0",
    "soupsieve==2.5",
    "spacy==3.7.1",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.4.8",
    "sympy==1.12",
    "tabulate==0.9.0",
    "thinc==8.2.1",
    "threadpoolctl==3.2.0",
    "tokenizers==0.14.0",
    "torch==2.1.0",
    "tqdm==4.66.1",
    "transformers==4.34.0",
    "trec-car-tools==2.6",
    "typer==0.9.0",
    "typing-extensions==4.8.0",
    "tzdata==2023.3",
    "unlzw3==0.2.2",
    "urllib3==2.0.6",
    "warc3-wet==0.2.3",
    "warc3-wet-clueweb09==0.2.5",
    "wasabi==1.1.2",
    "weasel==0.3.2",
    "wheel==0.37.1",
    "zipp==3.17.0",
    "zlib-state==0.1.6",
]
